plotting diffraction pattern now accepts Orientation class
