+++ py4DSTEM/process/diffraction/crystal_ACOM.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 
 from ...io.datastructure import PointList, PointListArray
 from ..utils import tqdmnd, electron_wavelength_angstrom
-# from . import Orientation
+from .utils import Orientation, OrientationMap, axisEqual3D
 
 
 def orientation_plan(
@@ -1169,19 +1169,30 @@ def match_single_pattern(
             ax[0].set_yticks([self.orientation_zone_axis_steps])
             ax[0].set_yticklabels([str(label_1)], size=14)
 
+
         # In-plane rotation
-        # ax[1].plot(
-        #     self.orientation_gamma * 180 / np.pi,
-        #     (np.squeeze(corr_full[ind_best_fit, :]) - cmin) / (cmax - cmin),
-        # )
         sig_in_plane = np.squeeze(corr_full[ind_best_fit, :])
+        sig_in_plane = sig_in_plane / np.max(sig_in_plane)
         ax[1].plot(
             self.orientation_gamma * 180 / np.pi,
-            sig_in_plane / np.max(sig_in_plane),
+            sig_in_plane,
         )
+
+        # Add markers for the best fit 
+        tol = 0.01
+        sub = sig_in_plane > 1 - tol
+        ax[1].scatter(
+                self.orientation_gamma[sub] * 180 / np.pi,
+                sig_in_plane[sub],
+                s=120,
+                linewidth=2,
+                facecolors="none",
+                edgecolors="r",
+            )
+
         ax[1].set_xlabel("In-plane rotation angle [deg]", size=16)
         ax[1].set_ylabel("Corr. of Best Fit Zone Axis", size=16)
-        ax[1].set_ylim([0, 1.01])
+        ax[1].set_ylim([0, 1.03])
 
         plt.show()
 
@@ -1193,86 +1204,6 @@ def match_single_pattern(
 
 
 
-
-
-@dataclass
-class Orientation:
-    """
-    A class for storing output orientations, generated by fitting a Crystal class orientation plan or
-    Bloch wave pattern matching to a PointList.
-
-    """
-    num_matches: int
-    matrix = None
-    corr = None
-    inds = None
-    mirror = None
-    angles = None
-
-    def __post_init__(self):
-        # initialize empty arrays
-        self.matrix = np.zeros((self.num_matches,3,3))
-        self.corr   = np.zeros((self.num_matches))
-        self.inds   = np.zeros((self.num_matches,2), dtype='int')
-        self.mirror = np.zeros((self.num_matches), dtype='bool')
-        self.angles = np.zeros((self.num_matches,3))
-
-@dataclass
-class OrientationMap:
-    """
-    A class for storing output orientations, generated by fitting a Crystal class orientation plan or
-    Bloch wave pattern matching to a PointListArray.
-
-    """
-    num_x: int
-    num_y: int
-    num_matches: int
-    matrix = None
-    corr = None
-    inds = None
-    angles = None
-
-    def __post_init__(self):
-        # initialize empty arrays
-        self.matrix = np.zeros((self.num_x,self.num_y,self.num_matches,3,3))
-        self.corr   = np.zeros((self.num_x,self.num_y,self.num_matches))
-        self.inds   = np.zeros((self.num_x,self.num_y,self.num_matches,2), dtype='int')
-        self.mirror = np.zeros((self.num_x,self.num_y,self.num_matches), dtype='bool')
-        self.angles = np.zeros((self.num_x,self.num_y,self.num_matches,3), dtype='float')
-
-
-    def set_orientation(self,orientation,ind_x,ind_y):
-        # Add an orientation to the orientation map
-        self.matrix[ind_x,ind_y] = orientation.matrix
-        self.corr[ind_x,ind_y] = orientation.corr
-        self.inds[ind_x,ind_y] = orientation.inds
-        self.mirror[ind_x,ind_y] = orientation.mirror
-        self.angles[ind_x,ind_y] = orientation.angles
-
-    def get_orientation(self,ind_x,ind_y):
-        # Return an orientation from the orientation map
-        orientation = Orientation(num_matches=self.num_matches)
-        orientation.matrix = self.matrix[ind_x,ind_y]
-        orientation.corr = self.corr[ind_x,ind_y]
-        orientation.inds = self.inds[ind_x,ind_y]
-        orientation.mirror = self.mirror[ind_x,ind_y]
-        orientation.angles = self.angles[ind_x,ind_y]
-
-        return orientation
-
-
-
-
-
-
-def axisEqual3D(ax):
-    extents = np.array([getattr(ax, "get_{}lim".format(dim))() for dim in "xyz"])
-    sz = extents[:, 1] - extents[:, 0]
-    centers = np.mean(extents, axis=1)
-    maxsize = max(abs(sz))
-    r = maxsize / 2
-    for ctr, dim in zip(centers, "xyz"):
-        getattr(ax, "set_{}lim".format(dim))(ctr - r, ctr + r)
 
 
 # zone axis range arguments for orientation_plan corresponding
@@ -1298,7 +1229,7 @@ orientation_ranges = {
     "-3": ["fiber", [0, 0, 1], [180.0, 60.0]],
     "32": ["fiber", [0, 0, 1], [90.0, 60.0]],
     "3m": ["fiber", [0, 0, 1], [180.0, 60.0]],
-    "-3m": ["fiber", [0, 0, 1], [180.0, 30.0]],
+    "-3m": ["fiber", [0, 0, 1], [90.0, 60.0]],
     "6": ["fiber", [0, 0, 1], [180.0, 60.0]],
     "-6": ["fiber", [0, 0, 1], [180.0, 60.0]],
     "6/m": [[[1, 0, 0], [0.5, np.sqrt(3) / 2.0, 0]], None, None],
